chore: move discourse comment endpoint to webhook routes
